refactor: NestedScrollConnection can be provided as a function
NestedScrollConnection can be provided as a function
